% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_infomap.R
\name{check_infomap}
\alias{check_infomap}
\title{Check if Infomap's standalone file is installed.}
\usage{
check_infomap(x = "Infomap")
}
\value{
An error if Infomap is not installed correctly, or version number of it is.
}
\description{
Attempts to run Infomap via the \code{system} command.
}
\details{
For now, this package depends on Infomap's stand-alone version. We are working on incorporating Infomap directly into the package.
}
