% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_monolayer_object.R
\name{create_monolayer_object}
\alias{create_monolayer_object}
\title{Creates a monolayer network object.}
\usage{
create_monolayer_object(
  x,
  directed = NULL,
  bipartite = NULL,
  group_names = c("set_cols", "set_rows"),
  node_metadata = NULL
)
}
\arguments{
\item{x}{An input list or matrix that must have node names.}

\item{directed}{Is the network directed?}

\item{bipartite}{Is the network bipartite?}

\item{group_names}{For bipartite networks: name of the groups in the columns and rows, respectively (e.g., parasites and hosts).}

\item{node_metadata}{Following the igraph method of \code{graph.data.frame}. Must have a column called node_name with names matching those in x.}
}
\value{
A \code{monolayer} object.
}
\description{
Automatically identifies if the input is a matrix or an edge list and creates a \code{monolayer} object.
Can also be used as a wrapper function for \code{list_to_matrix, matrix_to_list}
}
\details{
Converts between edge list and matrix formats as necessary.
}
\seealso{
\code{list_to_matrix, matrix_to_list, monolayer} functions and the \code{igraph} package.
}
