% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_multilayer_object.R
\name{create_multilayer_object}
\alias{create_multilayer_object}
\title{Create a multilayer network object}
\usage{
create_multilayer_object(
  extended,
  intra = NULL,
  inter = NULL,
  nodes = NULL,
  layers = NULL
)
}
\arguments{
\item{extended}{An extended edge list of the format \code{layer_from node_from layer_to node_to weight}, that contains all intra- and inter-layer edges. If supplied, then intra and inter are ignored. Should only contain node and layer IDs.}

\item{intra}{An edge list of the format \code{layer node_from node_to weight}. Should only contain node and layer IDs.}

\item{inter}{An edge list of the format \code{layer_from node_from layer_to node_to weight}. Should only contain node and layer IDs. Defaults to NULL, assuming the network does not have interlayer edges.}

\item{nodes}{A data frame with node metadata. First column must be node_id with IDs corresponding to those in \code{extended}, \code{intra} or \code{inter}.}

\item{layers}{A data frame with layer metadata. First column must be layer_id with IDs corresponding to those in \code{extended}, \code{intra} or \code{inter}.}
}
\value{
A \code{multilayer} object
}
\description{
The object includes the intralayer edges, interlayer edges (if there are), and metadata on nodes and layers.
The purpose of this package is to provide a wrapper for Infomap, and not to fully handle multilayer network objects. Therefore, converting and preparing the multilayer object is done by the user. A dedicated package for multilayer networks that handles data from scratch is in the works.
}
\seealso{
\code{multilayer}
}
