% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network_object.R
\name{create_network_object}
\alias{create_network_object}
\title{Convert between edge lists and matrices for monolayer networks}
\usage{
create_network_object(
  x,
  directed = NULL,
  bipartite = NULL,
  group_names = c("set_cols", "set_rows"),
  node_metadata = NULL
)
}
\arguments{
\item{x}{An input list or matrix that must have node names.}

\item{directed}{is the network directed}

\item{bipartite}{is network bipartite}

\item{group_names}{for bipartite networks: name of the groups in the columns and rows (e.g., parasites and hosts)}

\item{node_metadata}{Following the igraph method of \code{graph.data.frame}. Must have a column called node_name with names matching those in x.}
}
\value{
A \code{monolayer} object
}
\description{
Automatically identifies if the input is a matrix or edge list and creates a \code{monolayer} object.
A wrapper function for \code{list_to_matrix, matrix_to_list}
}
\seealso{
\code{list_to_matrix, matrix_to_list, monolayer}
}
