% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_color_hue.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Generate colors for ggplot}
\usage{
gg_color_hue(n, hue_min = 10, hue_max = 280, tune1 = 62, tune2 = 100)
}
\arguments{
\item{n}{See package metafolio}

\item{hue_min}{See package metafolio}

\item{hue_max}{See package metafolio}

\item{tune1}{See package metafolio}

\item{tune2}{See package metafolio}
}
\description{
Generate colors for ggplot
}
\details{
Taken from package metafolio. See details there
}
