% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_infomap.R
\name{install_infomap}
\alias{install_infomap}
\title{A wrapper to install Infomap's standalone file.}
\usage{
install_infomap(target_folder = NULL)
}
\arguments{
\item{target_folder}{Where to install infomap. Defaults to R's current working directory (\code{getwd()}).}
}
\value{
FALSE and an error if Infomap is not installed correctly. TRUE (and version number) if it is.
}
\description{
Clones the GitHub repository and runs \code{make} via R's \code{system} command. This is merely a wrapper for commands that can be easily run in the terminal.
}
\details{
For now, this package depends on Infomap's stand-alone version. Futre versions or another package will incorporate Infomap directly into R.
}
