% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kongsfjorden_nodes-data.R
\docType{data}
\name{kongsfjorden_nodes}
\alias{kongsfjorden_nodes}
\title{Kongsfjorden food web: nodes}
\format{
An object of class \code{data.frame} with 262 rows and 5 columns.
}
\usage{
data(kongsfjorden_nodes)
}
\description{
Food web including metazoan parasites for an intertidal ecosystem in New
Zealand.
}
\details{
The data set is included here "as is". In this network, nodes are
  species or taxonomical groups (e.g., Phytoplankton) and binary links
  represent feeding relationships. Kongsfjorden is a glacial fjord on the
  northwest corner of the Svalbard archipelago. It is a 30 km open fjord with
  no marked sill at the entrance, and with a maximum depth exceeding 300 m.
  The network consists of 262 species with 1,544 feeding interactions.
}
\examples{
data(kongsfjorden_nodes)
}
\references{
Jacob U, Thierry A, Brose U, Arntz WE, Berg S, Brey T, et al. The Role of Body Size in Complex Food Webs: A Cold Case. In: Belgrano A, editor. Advances in Ecological Research. Academic Press; 2011. pp. 181–223.

Cirtwill AR, Eklöf A. Feeding environment and other traits shape species’ roles in marine food webs. Ecol Lett. 2018;21: 875–884.
}
\seealso{
\code{kongsfjorden_links}
}
\keyword{datasets}
