% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_matrix.R
\name{list_to_matrix}
\alias{list_to_matrix}
\title{Edge list to matrix conversion}
\usage{
list_to_matrix(
  x,
  directed = F,
  bipartite = T,
  group_names = c("set_cols", "set_rows"),
  node_metadata = NULL
)
}
\arguments{
\item{x}{An edge list of the format from, to, weight.}

\item{directed}{Is the network directed?}

\item{bipartite}{Is network bipartite?}

\item{group_names}{For bipartite networks: name of the groups in the columns
and rows, respectively (e.g., parasites and hosts).}

\item{node_metadata}{Following the igraph method of \code{graph.data.frame}.
Must have a column called node_name with names matching those in x.}
}
\value{
A \code{monolayer} object.
}
\description{
Convert an edge list to a matrix. Can handle unipartite and bipartite
networks.
}
\details{
It is also possible to add node attributes. For this, it uses
igraph. The first column in the \code{node_metadata} data frame should have
the names of the nodes in the edge list.
}
\examples{
data(otago_nodes)
data(otago_links)

# Clean and prepare data
otago_nodes_2 <- otago_nodes \%>\%
filter(StageID==1) \%>\%
  select(node_name=WorkingName, node_id_original=NodeID, WorkingName,StageID, everything())
anyDuplicated(otago_nodes_2$node_name)
otago_links_2 <- otago_links \%>\%
filter(LinkType=='Predation') \%>\%
 filter(ConsumerSpecies.StageID==1) \%>\%
 filter(ResourceSpecies.StageID==1) \%>\%
 select(from=ResourceNodeID, to=ConsumerNodeID) \%>\%
 left_join(otago_nodes_2, by=c('from'='node_id_original')) \%>\%
 select(from, node_name, to) \%>\%
 left_join(otago_nodes_2, by=c('to'='node_id_original')) \%>\%
  select(from=node_name.x, to=node_name.y) \%>\%
 mutate(weight=1)
 
list_to_matrix(otago_links_2, directed = TRUE,
 bipartite = FALSE, node_metadata = otago_nodes_2)

}
\seealso{
Functions \code{create_monolayer_object, monolayer}, also,\code{graph.data.frame} from igraph.
}
