% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_list_bipartite.R
\name{matrix_to_list_bipartite}
\alias{matrix_to_list_bipartite}
\title{Incidence matrix to edge list conversion}
\usage{
matrix_to_list_bipartite(x, group_names = c("set_cols", "set_rows"))
}
\arguments{
\item{x}{An incidence matrix.}

\item{group_names}{Name of the groups in the columns and rows, respectively (e.g., parasites and hosts).}
}
\value{
A \code{monolayer} object.
}
\description{
Convert an incidence matrix to an edge list.
}
\details{
Assigns column and row names if those do not exist.
Cannot handle node attributes/metadata. Always assumes an undirected network.
}
\examples{
data(memmott1999)
matrix_to_list_bipartite(memmott1999, group_names = c('Animals', 'Plants'))

}
\seealso{
\code{create_monolayer_object, monolayer}
}
