% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilayer_alluvial.R
\name{multilayer_alluvial}
\alias{multilayer_alluvial}
\title{Alluvial plot for multilayer network}
\usage{
multilayer_alluvial(x)
}
\arguments{
\item{x}{An object of class \code{infomap_multilayer}.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Alluvial plot for multilayer network
}
\details{
Use ggplot and ggalluvial to plot an alluvial plot that shows the flow of nodes between modules. Very useful for temporal networks.
}
\seealso{
\code{ggplot2, ggalluvial, infomap_monolayer}
}
