% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multilayer_alluvial.R
\name{plot_multilayer_alluvial}
\alias{plot_multilayer_alluvial}
\title{Alluvial plot for multilayer network}
\usage{
plot_multilayer_alluvial(x, module_labels = F)
}
\arguments{
\item{x}{An object of class \code{infomap_multilayer}.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Alluvial plot for multilayer network
}
\details{
Use ggplot and ggalluvial to plot an alluvial plot that shows the flow of nodes between modules. Very useful for temporal networks. Because the output is
  a ggplot object it can be further manipulated with standard ggplot2.
}
\examples{
emln <- create_multilayer_object(extended = siberia1982_7_links, nodes = siberia1982_7_nodes, intra_output_extended = T, inter_output_extended = T)
emln_modules <- run_infomap_multilayer(M=emln, relax = F, flow_model = 'directed', silent = T, trials = 100, seed = 497294, temporal_network = T)

# Plot alluvial diagram
plot_multilayer_alluvial(emln_modules, module_labels = F)

# Can also modify the plot
plot_multilayer_alluvial(emln_modules, module_labels = F)+
   geom_stratum(linetype='dashed', color='gray')+
   scale_x_continuous(breaks=seq(0,6,1))+
   scale_y_continuous(breaks=seq(0,70,5))+
   labs(y='Number of species')+
   theme_bw()+
   theme(legend.position = "none",
         panel.grid = element_blank(),
         axis.text = element_text(color='black', size = 20),
         axis.title = element_text(size=20))
     
}
\seealso{
\code{ggplot2, ggalluvial, infomap_monolayer}
}
