% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signif.R
\name{plot_signif}
\alias{plot_signif}
\title{Plot histograms for randomizations}
\usage{
plot_signif(x, colors = c("plum", "tomato4"), plotit = T)
}
\arguments{
\item{x}{An object of class \code{infomap_monolayer}.}

\item{colors}{The fill colors of the histograms.}

\item{plotit}{Also show the plots?}
}
\value{
A list with objecst of class \code{ggplot}:
\itemize{
\item L_plot: Histogram for the map equation
\item m_plot: Histogram for the number of modules
}
}
\description{
Plot histograms for randomizations
}
\details{
Use ggplot to plot two histograms for the values of the map equation
 L and the number of modules. Dashed line is the observed value. Because
 returned objects are ggplot objects, they can be modified.
}
\examples{
network_object <- create_monolayer_object(bipartite::memmott1999, bipartite = T, directed = F, group_names = c('Animals','Plants'))
infomap_object <- run_infomap_monolayer(network_object, 
      infomap_executable='Infomap',
      flow_model = 'undirected',
      silent=T, trials=20, two_level=T, seed=123, 
      signif = T, shuff_method = 'r00', nsim = 50)

#get plots and plot them
x <- plot_signif(infomap_object, plotit = T)

# Can modify the plots with ggplot2
x$L_plot+
theme_bw()+
theme(legend.position='none', 
     axis.text = element_text(size=20), 
     axis.title = element_text(size=20))


}
\seealso{
\code{ggplot2, infomap_monolayer}
}
