% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_infomap.R
\name{shuffle_infomap}
\alias{shuffle_infomap}
\title{Shuffle a network and return linklists for Infomap (beta)}
\usage{
shuffle_infomap(x, shuff_method = "r00", nsim = 1000, ...)
}
\arguments{
\item{x}{An object of class \code{infomap_link_list} or \code{monolayer}.}

\item{shuff_method}{Method to shuffle the network. See details.}

\item{nsim}{How many shuffled networks to create?}

\item{...}{Additional parameters for shuffling passed to \code{simulate}, such as burnin. See \code{nullmodel} in package \code{vegan}.}
}
\value{
A list of linklists that can be used by Infomap.
}
\description{
Shuffles the network according to a predefined algorithm and return a list of
linklists that can be analyzed by Infomap.
}
\details{
This function is used internally by
  \code{run_infomap_monolayer}, but can also be run independently
  Currently only supports bipartite networks and shuffling methods from the
  vegan package as defined in \code{vegan::commsim}.
}
\examples{
data("memmott1999")
network_object <- create_monolayer_object(memmott1999, bipartite = T, directed = F, group_names = c('A','P'))
shuffled <- shuffle_infomap(network_object, shuff_method='curveball', nsim=50, burnin=2000)

}
\seealso{
Functions \code{commsim, nullmodel} in package \code{vegan}.
}
