% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siberia_interlayer.R
\docType{data}
\name{siberia_interlayer}
\alias{siberia_interlayer}
\title{Siberia interlayer edges}
\format{
Edge list format (layer_from node_from layer_to node_to weight).
}
\source{
Simulated data for illustration purposes.
}
\usage{
data(siberia_interlayer)
}
\description{
Host-parasite temporal multilayer network.
This data set contains the matrices of a temporal multilayer
network representing the infection of 22 small mammalian host species by
56 ectoparasite species during 6 consecutive summers in Siberia
(1982–1987).
}
\details{
The data set is in an matrix format.Each layer is a host-parasite bipartite network.
  Intralayer edges between a parasite species and a host species are the
  number of parasite individuals divided by the number of host individuals.
  Interlayer coupling edges connect each physical node to itself in the next
  layer (e.g., host A in layer 1 to host A in layer 2), and are calculated as
  the number of individuals in layer l+1 divided by the number of individuals
  in layer l. They therefore represent population dynamics. Interlayer edges
  only go one way (_l-->l+1_) because time flow one way. 
  
  See \code{siberia_nodes} for information about the nodes.
  
  See \code{siberia1982_matrix} Siberia 1982 host-parasite temporal matrix.
}
\examples{
data(siberia1982_matrix)
data(siberia1983_matrix)
data(siberia1984_matrix)
data(siberia1985_matrix)
data(siberia1986_matrix)
data(siberia1987_matrix)
data(siberia_nodes)
data(siberia_interlayer)

# Create a multilayer object
layer_attrib <- tibble(layer_id=1:6, layer_name=c('1982','1983','1984','1985','1986','1987'))                                     

multilayer_siberia <- create_multilayer_network(list_of_layers = list(siberia1982_matrix,
                                                                      siberia1983_matrix,
                                                                      siberia1984_matrix,
                                                                      siberia1985_matrix,
                                                                      siberia1986_matrix,
                                                                      siberia1987_matrix),
                                                layer_attributes = layer_attrib,
                                                interlayer_links = siberia_interlayer,
                                                bipartite = T,
                                                directed = F, physical_node_attributes = siberia_nodes )
}
\references{
Krasnov BR, Matthee S, Lareschi M, Korallo-Vinarskaya NP, Vinarski MV. Co-occurrence of ectoparasites on rodent hosts: null model analyses of data from three continents. Oikos. 2010;119: 120–128.

Pilosof S, Fortuna MA, Vinarski MV, Korallo-Vinarskaya NP, Krasnov BR. Temporal dynamics of direct reciprocal and indirect effects in a host-parasite network. J Anim Ecol. 2013;82: 987–996.

Pilosof S, Porter MA, Pascual M, Kéfi S. The multilayer nature of ecological networks. Nature Ecology & Evolution. 2017;1: 0101.
}
\keyword{datasets}
