% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_infomap.R
\name{check_infomap}
\alias{check_infomap}
\title{Check if Infomap's standalone file is installed.}
\usage{
check_infomap(x = "Infomap")
}
\arguments{
\item{x}{The name of the Infomap executable file.}
}
\value{
FALSE and an error if Infomap is not installed correctly. TRUE (and version number) if it is.
}
\description{
Attempts to run Infomap via the \code{system} command to make sure it works propoerly.
}
\details{
For now, this package depends on Infomap's stand-alone version. Futre versions or another package will incorporate Infomap directly into R.
}
\examples{
check_infomap()

}
