% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_monolayer_object.R
\name{create_monolayer_object}
\alias{create_monolayer_object}
\title{Creates a monolayer network object.}
\usage{
create_monolayer_object(
  x,
  directed = NULL,
  bipartite = NULL,
  group_names = c("set_cols", "set_rows"),
  node_metadata = NULL
)
}
\arguments{
\item{x}{input data in the format of a matrix, edge list or an igraph object.}

\item{directed}{Is the network directed?}

\item{bipartite}{Is the network bipartite?}

\item{group_names}{For bipartite networks: name of the groups in the columns
and rows, respectively (e.g., parasites and hosts).}

\item{node_metadata}{Following the igraph method of \code{graph.data.frame}.
Must have a column called node_name with names matching those in x.}
}
\value{
A \code{monolayer} object.
}
\description{
Automatically identifies if the input is a matrix or an edge list and creates
a \code{monolayer} object.
}
\details{
Converts between edge list and matrix formats and creates a
  monolayer object. It a wrapper function for
  \code{list_to_matrix, matrix_to_list_unipartite, matrix_to_list_bipartite}. Node metadata can only be included with an edge list input.
}
\examples{
x <- create_monolayer_object(memmott1999, bipartite = T, directed = F, group_names = c('A','P'))

}
\seealso{
\code{list_to_matrix, matrix_to_list_unipartite,
  , monolayer}. Also functions \code{graph.incidence,
  graph.adjacency, graph.data.frame} in the \code{igraph} package.
}
