% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_multilayer_object.R
\name{create_multilayer_object}
\alias{create_multilayer_object}
\title{Create a multilayer network object}
\usage{
create_multilayer_object(
  extended = NULL,
  intra = NULL,
  inter = NULL,
  nodes = NULL,
  intra_output_extended = T,
  layers = NULL,
  write_to_file = F,
  filename_prefix = NULL
)
}
\arguments{
\item{extended}{An extended edge list input of the format \code{layer_from
node_from layer_to node_to weight}, that contains all intralayer, and
(optionally) interlayer edges. If supplied, then \code{intra} and
\code{inter} are ignored. Nodes and layers should be specified by IDs, not names.}

\item{intra}{An edge list of the format \code{layer node_from node_to
weight}. Nodes and layers should be specified by IDs, not names.}

\item{inter}{An extended edge list of the format \code{layer_from node_from
layer_to node_to weight}. Nodes and layers should be specified by IDs, not names. Defaults to NULL,
 assuming the network does not have interlayer edges.}

\item{nodes}{A data frame with node metadata. First column must be node_id
with IDs corresponding to those in \code{extended}, \code{intra} or
\code{inter}.}

\item{intra_output_extended}{Defaults to \code{TRUE}. Should the intralayer edges be written in the extended edge list format. 
Set as \code{FALSE} for running with relax_rate instead of with interlayer edges.}

\item{layers}{A data frame with layer metadata. First column must be layer_id
with IDs corresponding to those in \code{extended}, \code{intra} or
\code{inter}.}

\item{write_to_file}{Write the edge list to a text file? Convenient to copy-paste into the online version of infomap.}

\item{filename_prefix}{The name of the file. Extension will always be .txt and the file is space-delimited.}
}
\value{
A \code{multilayer} object
}
\description{
Takes an edge list input with node IDs and returns a multilayer object that
includes the intralayer edges, interlayer edges (if there are), and metadata
on nodes and layers.
}
\details{
\code{extended, intra, inter} must only contain node IDs. Any
  node and layer attributes that link to these IDs should be provided in
  \code{nodes} and \code{layers}.

  Intralayer edges can have either an extended format of \code{layer_from
  node_from layer_to node_to weight} or a non-extended format of \code{layer
  node_from node_to weight}.

  Interlayer edges must have an extended format of \code{layer_from
  node_from layer_to node_to weight}. This is because a non-extended format (\code{layer_from
  node_from layer_to weight}) is ONLY valid for interlayer edges that connect between different
  realizations of the same physical nodes in different layers (coupling interlayer links). To avoid
  confusion we do not allow this format.
}
\examples{
data("siberia1982_7_links") # Links are an extended edge list
data("siberia1982_7_nodes") # nodes

# Information on layers
layers <- tibble(layer_id=1:6, year=1982:1987)

# Create a multilayer object with an extended list.
emln <- create_multilayer_object(extended = siberia1982_7_links,
 nodes = siberia1982_7_nodes,
 intra_output_extended = TRUE,
 layers=layers)

# Create a multilayer object with a non-extended Intra/Inter type link list.
emln <- create_multilayer_object(extended = siberia1982_7_links,
 nodes = siberia1982_7_nodes,
 intra_output_extended = FALSE,
 layers=layers)

# To ignore interlayer edges:
emln$inter <- NULL


}
\seealso{
\code{multilayer}
}
