% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infomapecology.R
\docType{package}
\name{infomapecology}
\alias{infomapecology}
\title{infomapecology: A package for running Infomap with ecological networks}
\description{
Wrapper functions to run Infomap. Inspired by ecological data.
Also includes several data sets
}
\references{
A dynamical perspective...
}
