% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infomapecology.R
\docType{package}
\name{infomapecology}
\alias{infomapecology}
\title{infomapecology: A package for running Infomap, inspired by ecological networks}
\description{
The package contains functions that prepare monolayer and multilayer networks
for analysis with Infomap, run Infomap from within R and parse the results
back to R. Also includes standardized objects to store and further analyze
monolayer and multilayer networks. The package is inspired by analysis of
ecological networks but is suitable for other areas of research as well.
}
\details{
We use Infomap as a stand-alone file. Future versions will integrate Infomap
directly into R. Complete instructions on how to download, install and use
Infomap can be found in
\href{https://www.mapequation.org}{https://www.mapequation.org}. The best
practice is to compile Infomap under the file name Infomap and place it in the
same working folder in which the R code is run. See details in \href{https://ecological-complexity-lab.github.io/infomap_ecology_package/installation}{https://ecological-complexity-lab.github.io/infomap_ecology_package/installation}

The package also includes several data sets of ecological networks.

Ideas/requests for features are welcome (open an issue
\href{https://github.com/Ecological-Complexity-Lab/infomap_ecology_package/issues}{https://github.com/Ecological-Complexity-Lab/infomap_ecology_package/issues}).
}
\references{
Farage C, Edler D, Eklöf A, Rosvall M, Pilosof S. Identifying flow
 modules in ecological networks using Infomap. Methods in Ecology and
 Evolution 2021. doi:10.1111/2041‐210X.13569
}
\author{
Shai Pilosof
}
