% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infomapecology.R
\docType{package}
\name{infomapecology}
\alias{infomapecology}
\title{infomapecology: A package for running Infomap, inspired by ecological networks}
\description{
The package contains functions that prepare monolayer and multilayer networks
for analysis with Infomap, run Infomap from within R and parse the results
back to R. Also includes standardized objects to store and further analyze
monolayer and multilayer networks. The package is inspired by analysis of
ecological networks but is suitable for other areas of research as well.
}
\details{
We use Infomap version 1 as a stand-alone file. Future versions (or a differen
package) will integrate Infomap directly intro R. Complete instructions on how
to download, install and use Infomap can be found in
\href{https://www.mapequation.org}{https://www.mapequation.org}. The best
practice is to compile Infomap under the file name Infomap and place it in the
same working folder in which the R code is run.

The package also includes several data sets of ecological networks.

Inspired by ecological networks but can work for other networks too!
Ideas/requests for features are welcome (open an issue \href{here}{https://github.com/Ecological-Complexity-Lab/infomap_ecology_package/issues}).
}
\references{
A dynamical perspective to community detection in ecological
 networks.
}
\author{
Shai Pilosof
}
