% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_matrix.R
\name{list_to_matrix}
\alias{list_to_matrix}
\title{Edge list to matrix conversion}
\usage{
list_to_matrix(
  x,
  directed = F,
  bipartite = T,
  group_names = c("set_cols", "set_rows"),
  node_metadata = NULL
)
}
\arguments{
\item{x}{An edge list of the format from, to, weight}

\item{directed}{is the network directed}

\item{bipartite}{is network bipartite}

\item{group_names}{for bipartite networks: name of the groups in the columns
and rows (e.g., parasites and hosts)}

\item{node_metadata}{Following the igraph method of \code{graph.data.frame}.
Must have a column called node_name with names matching those in x.}
}
\value{
A \code{monolayer} object
}
\description{
Convert an edge list to a matrix. Can handle unipartite and bipartite
networks. It is also possible to add node attributes. For this, it uses
igraph. The first column in the\code{node_metadata} data frame should have
the names of the nodes in the edge list.
}
\seealso{
\code{create_network_object, monolayer}
}
