% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_modular_matrix.R
\name{plot_modular_matrix}
\alias{plot_modular_matrix}
\title{Plot a matrix with colored interactions by modules}
\usage{
plot_modular_matrix(
  x,
  fix_coordinates = T,
  axes_titles = c("Set 1", "Set 2"),
  transpose = F,
  outside_module_col = "gray"
)
}
\arguments{
\item{x}{An object of class \code{infomap_monolayer}.}

\item{fix_coordinates}{Should x and y coordinates be fixed by ggplot to
create square cells.}

\item{transpose}{Option to transpose the matrix. Useful for matrices with
very differnt numbers of rows and columns.}

\item{outside_module_col}{Color of interactions outside modules.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Plot a matrix with colored interactions by modules
}
\details{
Use ggplot to plot a matrix in which cells (network interactions)
  are colored. Interactions that fall outside the modules are colored in
  \code{outside_module_col} and those that fall inside the module are colored
  by colors automativaclly generated by \code{gg_color_hue} from package metafolio:
  
  \code{gg_color_hue <- function(n, hue_min = 10, hue_max = 280, tune1 = 62, tune2 = 100) {
hues = seq(hue_min, hue_max, length=n+1)
hcl(h=hues, l=tune1, c=tune2)[1:n]}
}
}
\seealso{
\code{ggplot2, infomap_monolayer}
}
