% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_infomap_monolayer.R
\name{run_infomap_monolayer}
\alias{run_infomap_monolayer}
\title{Run Infomap for monolayer networks}
\usage{
run_infomap_monolayer(
  x,
  infomap_executable = "Infomap",
  input_file = NULL,
  flow_model = NULL,
  silent = F,
  trials = 20,
  two_level = T,
  seed = 123,
  ...
)
}
\arguments{
\item{x}{A monolayer object. Ignored if an input file is provided.}

\item{infomap_executable}{Name of Infomap standalone file (default is Infomap).}

\item{input_file}{Name of an external input file. Overrides x}

\item{flow_model}{See details in https://www.mapequation.org/infomap/#ParamsAlgorithm.}

\item{silent}{Run in silent mode (argumnt --silent in Infomap)}

\item{trials}{Number of trials to run (argumnt -N in Infomap)}

\item{two_level}{Do not use hierarchical partitioning (argumnt -2 in Infomap)}

\item{seed}{Seed value for random number generation (argumnt --seed in Infomap)}

\item{...}{additional Infomap arguments as detailed in https://www.mapequation.org/infomap/#Parameters.}
}
\value{
A list:
\itemize{
  \item \code{L} The value of the map equation
  \item \code{modules} A tibble with node IDs. For each node, the number of module levels (e.g., 2 levels if it belongs to module 3 within module 2), and module affiliation, by levels.
  \item \code{module_affiliation} A tibble with node id, name, module affiliations and node attributes (metadata).
}
}
\description{
Run Infomap for monolayer networks using a \code{monolayer} class object, or
using an external file. Returns the value of the map equation and a tibble
with modules, If using a \code{monolayer} object, also returns a tibble with
module affiliations that includes additional node metadata (if exists).
}
\seealso{
\code{create_network_object, monolayer}
}
